+++ easydel/modules/deepseek_v2/modeling_deepseek_flax.py
@@ -823,8 +823,8 @@ class DeepseekV2Model(EasyDeLBaseModule):
 	def __init__(
 		self,
 		config: DeepseekV2Config,
-		dtype: jnp.dtype = jnp.float32,
-		param_dtype: jnp.dtype = jnp.float32,
+		dtype: jnp.dtype = jnp.bfloat16,
+		param_dtype: jnp.dtype = jnp.bfloat16,
 		precision: jax.lax.PrecisionLike = None,
 		*,
 		rngs: nn.Rngs,
@@ -1019,8 +1019,8 @@ class DeepseekV2ForCausalLM(EasyDeLBaseModule):
 	def __init__(
 		self,
 		config: DeepseekV2Config,
-		dtype: jnp.dtype = jnp.float32,
-		param_dtype: jnp.dtype = jnp.float32,
+		dtype: jnp.dtype = jnp.bfloat16,
+		param_dtype: jnp.dtype = jnp.bfloat16,
 		precision: jax.lax.PrecisionLike = None,
 		*,
 		rngs: nn.Rngs,
